% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/now_serving.R
\name{now_serving}
\alias{now_serving}
\title{Now Serving...}
\usage{
now_serving(n = 1)
}
\arguments{
\item{n}{the number of emojis to return. Default is 1.}
}
\value{
A character vector.
}
\description{
This is a function which delivers a random sampling of food emojis
}
\examples{
now_serving()

now_serving(n=3)
}
